// src/app/models/drone.model.ts
export class Drone {
  /*
  id?: any;
  title?: string;
  description?: string;
  published?: boolean;
  */
  id?: any;
  title?: string;
  brand?: string;
  price?: any;
  asinid?: string;
  rating?: any;
  stockqnty?: any;
  amazon_image?: string;
  amazon_url?: string;
  synopsis?: string;
  description?: string;
  published?: boolean;
}